- Fixed borked refactoring in r307437 (using SUCCESS/FAILURE return instead of   out parameter). - Fixed signature of php_stream_copy_to_stream_ex to return int in 5.4/trunk   instead of size_t, as the function only returns SUCCESS/FAILURE.
git-svn-id: http://svn.php.net/repository/php/php-src/trunk@316121 c90b9560-bf6c-de11-be94-00142212c4b1

+++ main/php_streams.h
@@ -438,7 +438,7 @@ BEGIN_EXTERN_C()
 ZEND_ATTRIBUTE_DEPRECATED
 PHPAPI size_t _php_stream_copy_to_stream(php_stream *src, php_stream *dest, size_t maxlen STREAMS_DC TSRMLS_DC);
 #define php_stream_copy_to_stream(src, dest, maxlen)	_php_stream_copy_to_stream((src), (dest), (maxlen) STREAMS_CC TSRMLS_CC)
-PHPAPI size_t _php_stream_copy_to_stream_ex(php_stream *src, php_stream *dest, size_t maxlen, size_t *len STREAMS_DC TSRMLS_DC);
+PHPAPI int _php_stream_copy_to_stream_ex(php_stream *src, php_stream *dest, size_t maxlen, size_t *len STREAMS_DC TSRMLS_DC);
 #define php_stream_copy_to_stream_ex(src, dest, maxlen, len)	_php_stream_copy_to_stream_ex((src), (dest), (maxlen), (len) STREAMS_CC TSRMLS_CC)
